+++ mf2_str_defs.h
@@ -17,6 +17,7 @@
 #define MF2_STR_DEFS \
 	X(items) \
 	X(rels) \
+	X(relurls) \
 	X(rel_urls) \
 
 #endif /* MF2_STR_DEFS_H */